+++ src/Data/ChromaAppInfo.cs
@@ -9,7 +9,7 @@ namespace ChromaWrapper.Data
     /// <seealso href="https://assets.razerzone.com/dev_portal/C%2B%2B/html/en/struct_chroma_s_d_k_1_1_a_p_p_i_n_f_o_t_y_p_e.html">ChromaSDK::APPINFOTYPE Struct Reference</seealso>
     [SuppressMessage("Style", "IDE0032:Use auto property", Justification = "Interop marshaling")]
     [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Unicode)]
-    public sealed class ChromaAppInfo
+    public sealed class ChromaAppInfo : IChromaAppInfo
     {
         [MarshalAs(UnmanagedType.ByValTStr, SizeConst = 256)]
         private string? _title;